% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sub_matrix.R
\name{build_sub_matrix}
\alias{build_sub_matrix}
\title{Build a Substitution Matrix}
\usage{
build_sub_matrix(simple_version = FALSE, max_len_score = 4L)
}
\arguments{
\item{simple_version}{if \code{TRUE}, just use segmental copy number value.}

\item{max_len_score}{the maximum score for segment length (should >=4).
The maximum score for copy number value is 6 (cannot be changed).}
}
\value{
a \code{list}.
}
\description{
Build a Substitution Matrix
}
\examples{
sub_list <- build_sub_matrix()
sub_list2 <- build_sub_matrix(simple_version = TRUE)
}
