% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_pam.R
\name{cluster_pam_estimate}
\alias{cluster_pam_estimate}
\title{Estimate Optimal Number of Cluster for PAM Algorithm}
\usage{
cluster_pam_estimate(
  x,
  method = c("silhouette", "wss", "gap_stat"),
  k.max = 10,
  nboot = 100,
  verbose = interactive(),
  barfill = "steelblue",
  barcolor = "steelblue",
  linecolor = "steelblue",
  print.summary = TRUE,
  FUNcluster = cluster::pam,
  seed = 1234L,
  ...
)
}
\arguments{
\item{x}{a dissimilarity matrix.}

\item{method}{the method to be used for estimating the optimal number of 
clusters. Possible values are "silhouette" (for average silhouette width), 
"wss" (for total within sum of square) and "gap_stat" (for gap statistics).}

\item{k.max}{the maximum number of clusters to consider, must be at least two.}

\item{nboot}{integer, number of Monte Carlo ("bootstrap") samples. Used only for determining the number of clusters 
using gap statistic.}

\item{verbose}{logical value. If TRUE, the result of progress is printed.}

\item{barfill}{fill color and outline color for bars}

\item{barcolor}{fill color and outline color for bars}

\item{linecolor}{color for lines}

\item{print.summary}{logical value. If true, the optimal number of clusters 
are printed in fviz_nbclust().}

\item{FUNcluster}{a partitioning function which accepts as first argument a 
(data) matrix like x, second argument, say k, k >= 2, the number of 
clusters desired, and returns a list with a component named cluster which 
contains the grouping of observations. Allowed values include: kmeans,
cluster::pam, cluster::clara, cluster::fanny, hcut, etc. This argument is
not required when x is an output of the function 
\code{NbClust::NbClust()}.}

\item{...}{optionally further arguments for FUNcluster()}
}
\value{
a \code{ggplot} object.
}
\description{
Estimate Optimal Number of Cluster for PAM Algorithm
}
