% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_split.R
\name{cluster_split}
\alias{cluster_split}
\title{Split Cluster Sequence into List}
\usage{
cluster_split(x, s = NULL, block_size = 10)
}
\arguments{
\item{x}{a named integer vector from \code{hclust} etc.}

\item{s}{default is \code{NULL}, when the \code{x} is block cluster sequence, set this
to a sequence vector.}

\item{block_size}{block size used to split, only used when \code{s} is not \code{NULL}.}
}
\value{
a \code{list}.
}
\description{
Split Cluster Sequence into List
}
