% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMotif-package.R
\docType{package}
\name{CNVMotif-package}
\alias{CNVMotif}
\alias{CNVMotif-package}
\title{CNVMotif: Explore, Analyze and Visualize Catalogs and Patterns of
    Copy Number Variation in Cancer Genomics}
\description{
Provides functionality for exploring, analyzing and
    visualizing the copy number variation (CNV) motifs in cancer genomics.
}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
