% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_msa.R
\name{do_msa}
\alias{do_msa}
\title{Run Modified Multiple Sequence Alignment}
\usage{
do_msa(
  x,
  substitutionMatrix = NULL,
  gapOpening = 6,
  gapExtension = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a character vector.}

\item{substitutionMatrix}{substitution matrix for scoring matches and mismatches.
Default is \code{NULL}, use \code{1} for match and \code{0} for unmatch.}

\item{gapOpening}{gap opening penalty; Note that the sign of this parameter is ignored.}

\item{gapExtension}{gap extension penalty; Note that the sign of this parameter is ignored.}

\item{verbose}{if \code{TRUE}, print extra info.}

\item{...}{other arguments passing to \link[msa:msa]{msa::msa}}
}
\value{
a \code{list}.
}
\description{
Run Modified Multiple Sequence Alignment
}
\examples{
r <- do_msa(c("ABCDF", "BCDEF"))
r
}
