% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs.R
\name{extract_seqs}
\alias{extract_seqs}
\title{Extract Pasted Sequences from Each Chromosome}
\usage{
extract_seqs(
  dt,
  len = 5L,
  step = 1L,
  local_cutoff = 1e+07,
  flexible_approach = FALSE,
  return_dt = FALSE
)
}
\arguments{
\item{dt}{a \code{data.table} from \link{transform_seqs}.}

\item{len}{cut length.}

\item{step}{step size to move on each chromosome sequence.}

\item{local_cutoff}{any segment with length greater than this cutoff will be filtered out and
used as cutpoint, default is \verb{10Mb}.}

\item{flexible_approach}{if \code{TRUE}, extract flexible-size sequences between segments
with size less than specified cutoff. So the arguments \code{len} and \code{step} are ignored.}

\item{return_dt}{if \code{TRUE}, just return a \code{data.table} containing
mutated \code{Seqs} column.}
}
\value{
a \code{list}.
}
\description{
See \code{\link[=get_score_matrix]{get_score_matrix()}} for examples.
The result sequences are unique and sorted.
}
