% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlogo.R
\name{ggseqlogo2}
\alias{ggseqlogo2}
\alias{geom_logo2}
\title{Quick sequence logo plot}
\usage{
ggseqlogo2(
  data,
  facet = "wrap",
  scales = "free_x",
  ncol = NULL,
  nrow = NULL,
  idor = NULL,
  ...
)

geom_logo2(
  data = NULL,
  method = "bits",
  seq_type = "auto",
  namespace = NULL,
  font = "roboto_medium",
  stack_width = 0.95,
  rev_stack_order = F,
  col_scheme = "auto",
  low_col = "black",
  high_col = "yellow",
  na_col = "grey20",
  plot = TRUE,
  idor = NULL,
  ...
)
}
\arguments{
\item{data}{Character vector of sequences or named list of sequences. All sequences must have same width}

\item{facet}{Facet type, can be 'wrap' or 'grid'}

\item{scales}{Facet scales, see \code{\link{facet_wrap}}}

\item{ncol}{Number of columns, works only when \code{facet='wrap'}, see \code{\link{facet_wrap}}}

\item{nrow}{Number of rows, same as \code{ncol}}

\item{idor}{a named vector (like a dictory) to change letters one to one in the plot.}

\item{...}{Additional arguments passed to \code{\link[ggseqlogo]{geom_logo}}}

\item{method}{Height method, can be one of "bits" or "probability" (default: "bits")}

\item{seq_type}{Sequence type, can be one of "auto", "aa", "dna", "rna" or "other" 
(default: "auto", sequence type is automatically guessed)}

\item{namespace}{Character vector of single letters to be used for custom namespaces. Can be alphanumeric, including Greek characters.}

\item{font}{Name of font. See \code{list_fonts} for available fonts.}

\item{stack_width}{Width of letter stack between 0 and 1 (default: 0.95)}

\item{rev_stack_order}{If \code{TRUE}, order of letter stack is reversed (default: FALSE)}

\item{col_scheme}{Color scheme applied to the sequence logo. See \code{list_col_schemes} for available fonts.
(default: "auto", color scheme is automatically picked based on \code{seq_type}). 
One can also pass custom color scheme objects created with the \code{make_col_scheme} function}

\item{low_col}{Colors for low and high ends of the gradient if a quantitative color scheme is used (default: "black" and "yellow").}

\item{high_col}{Colors for low and high ends of the gradient if a quantitative color scheme is used (default: "black" and "yellow").}

\item{na_col}{Color for letters missing in color scheme (default: "grey20")}

\item{plot}{If \code{FALSE}, plotting data is returned}
}
\description{
\code{ggseqlogo} is a shortcut for generating sequence logos. 
It adds the ggseqlogo theme \code{\link[ggseqlogo]{theme_logo}} by default, and facets when multiple input data are provided. 
It serves as a convenient wrapper, so to customise logos beyond the defaults here, please use \code{\link[ggseqlogo]{geom_logo}}.
}
\examples{
library(ggseqlogo)
data(ggseqlogo_sample)

## Same as ggseqlogo()
p1 <- ggseqlogo2(seqs_dna[[1]])
p1

## Extra feature
idor <- as.character(1:4)
names(idor) <- c("A","C","G","T")
p2 <- ggseqlogo2(seqs_dna[[1]], idor = idor)
p2

}
