% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_motif_finder.R
\name{sh_code_segs}
\alias{sh_code_segs}
\title{Coding Copy Number Segments with Letters.}
\usage{
sh_code_segs(x)
}
\arguments{
\item{x}{a \code{CopyNumber} object or a \code{data.frame}.}
}
\value{
a \code{data.table}.
}
\description{
Letters are grouped as short (<50kb), mid (<500kb), long (<5Mb),
long (or extreme) long (>5Mb) segments.
\itemize{
\item A B C D for copy number 0.
\item E F G H for copy number 1.
\item I J K L for copy number 2.
\item M N O P for copy number 3.
\item Q R S T for copy number 4.
\item U V W X for copy number 5+.
}
}
