% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_motif_finder.R
\name{show_seq_shape}
\alias{show_seq_shape}
\title{Show Copy Number Sequence Shapes}
\usage{
show_seq_shape(
  x,
  map = NULL,
  x_lab = "Estimated segment length",
  y_lab = "Copy number"
)
}
\arguments{
\item{x}{a character vector of sequences or named list of sequences. All sequences must have same width.}

\item{map}{default is \code{NULL}, a named vector.}

\item{x_lab}{x lab.}

\item{y_lab}{y lab.}
}
\value{
a \code{ggplot} object.
}
\description{
Show Copy Number Sequence Shapes
}
\examples{
p <- show_seq_shape(c("ADGHK"))
p
}
