% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_pam.R
\name{cluster_pam}
\alias{cluster_pam}
\title{PAM Wrapper}
\usage{
cluster_pam(x, k, ...)
}
\arguments{
\item{x}{a dissimilarity matrix.}

\item{k}{positive integer specifying the number of clusters, less than
    the number of observations.}

\item{...}{other parameters passing to \link[cluster:pam]{cluster::pam}.}
}
\value{
a PAM clustering result object.
}
\description{
See \code{\link[factoextra:fviz_cluster]{factoextra::fviz_cluster()}} for visualization.
}
