% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_motif_finder.R
\name{sh_build_sub_matrix}
\alias{sh_build_sub_matrix}
\title{Build a Substitution Matrix}
\usage{
sh_build_sub_matrix(max_len_score = 8L)
}
\arguments{
\item{max_len_score}{the maximum score for segment length (should >=4).
Default is 8 for balancing the weights between segment length and copy number
value. The maximum score for copy number value is 6.}
}
\value{
a \code{list}.
}
\description{
Build a Substitution Matrix
}
\examples{
sub_list <- sh_build_sub_matrix
}
