% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_motif_finder.R
\name{sh_coding_segs}
\alias{sh_coding_segs}
\title{Coding Copy Number Segments with Letters.}
\usage{
sh_coding_segs(x, max_len_score = 8L)
}
\arguments{
\item{x}{a \code{CopyNumber} object or a \code{data.frame} with
at least 5 columns ("sample", "chromosome", "start", "end", "segVal").}

\item{max_len_score}{the maximum score for segment length (should >=4).
Default is 8 for balancing the weights between segment length and copy number
value. The maximum score for copy number value is 6.}
}
\value{
a \code{list}.
}
\description{
See \code{\link[=sh_get_score_matrix]{sh_get_score_matrix()}} for examples.
Letters are grouped as short (<50kb), mid (<500kb), long (<5Mb),
long (or extreme) long (>5Mb) segments.
\itemize{
\item A B C D for copy number 0.
\item E F G H for copy number 1.
\item I J K L for copy number 2.
\item M N O P for copy number 3.
\item Q R S T for copy number 4.
\item U V W X for copy number 5+.
}
}
