% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_motif_finder.R
\name{sh_extract_seqs}
\alias{sh_extract_seqs}
\title{Extract Pasted Sequences from Each Chromosome}
\usage{
sh_extract_seqs(dt, len = 5L, step = 2L, return_dt = FALSE)
}
\arguments{
\item{dt}{a \code{data.table} from \link{sh_coding_segs}.}

\item{len}{cut length.}

\item{step}{step size to move on each chromosome sequence.}

\item{return_dt}{if \code{TRUE}, just return a \code{data.table} containing
mutated \code{Seqs} column.}
}
\value{
a \code{list}.
}
\description{
See \code{\link[=sh_get_score_matrix]{sh_get_score_matrix()}} for examples.
The result sequences are unique and sorted.
}
