% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_seq_logo.R
\name{show_seq_logo}
\alias{show_seq_logo}
\title{Show Copy Number Sequence Logos}
\usage{
show_seq_logo(
  x,
  method = c("prob", "bits"),
  simple_version = FALSE,
  ncol = NULL,
  nrow = NULL,
  recode = FALSE,
  indicator = NULL,
  ...
)
}
\arguments{
\item{x}{a character vector of sequences or named list of sequences. All sequences must have same width.}

\item{method}{Height method, can be one of "bits" or "probability" (default: "bits")}

\item{simple_version}{if \code{TRUE}, just use segmental copy number value.}

\item{ncol}{Number of columns, works only when \code{facet='wrap'}, see \code{\link{facet_wrap}}}

\item{nrow}{Number of rows, same as \code{ncol}}

\item{recode}{if \code{TRUE}, it will use default indicator or specified indicator to show the letters in
the plot}

\item{indicator}{a named vector (like a dictory) to change letters one to one in the plot.}

\item{...}{Additional arguments passed to \code{\link[ggseqlogo]{geom_logo}}}
}
\value{
a \code{ggplot} object
}
\description{
Show Copy Number Sequence Logos
}
\examples{
p1 <- show_seq_logo(sapply(split(LETTERS[1:24], 1:4), function(x) paste0(x, collapse = "")))
p1
p2 <- show_seq_logo(sapply(split(LETTERS[1:24], 1:4), function(x) paste0(x, collapse = "")),
  recode = TRUE
)
p2
p3 <- show_seq_logo(sapply(split(LETTERS[1:6], 1:2), function(x) paste0(x, collapse = "")),
  simple_version = TRUE
)
}
