% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_seq_shape.R
\name{show_seq_shape}
\alias{show_seq_shape}
\title{Show Copy Number Sequence Shapes}
\usage{
show_seq_shape(
  x,
  map = NULL,
  simple_version = FALSE,
  line_size_scale = 3,
  x_lab = ifelse(simple_version, "Assumed equal length", "Estimated segment length"),
  y_lab = "Copy number",
  nrow = NULL,
  ncol = NULL,
  scales = "free_x"
)
}
\arguments{
\item{x}{a character vector of sequences or named list of sequences. All sequences must have same width.}

\item{map}{default is \code{NULL}, a named string vector.}

\item{simple_version}{if \code{TRUE}, just use segmental copy number value.}

\item{line_size_scale}{the scale size for line width.}

\item{x_lab}{x lab.}

\item{y_lab}{y lab.}

\item{nrow}{Number of rows, same as \code{ncol}}

\item{ncol}{Number of columns, works only when \code{facet='wrap'}, see \code{\link{facet_wrap}}}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}
}
\value{
a \code{ggplot} object.
}
\description{
Show Copy Number Sequence Shapes
}
\examples{
p <- show_seq_shape(c("ADGHK"))
p
x <- list(a = c("ABCDE", "AXFDP"), b = c("KKDFH", "GKDFM"))
p2 <- show_seq_shape(x)
p2
}
