% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminer.helper-package.R
\docType{package}
\name{sigminer.helper-package}
\alias{sigminer.helper}
\alias{sigminer.helper-package}
\title{sigminer.helper: Extra Functionality for Package 'sigminer'}
\description{
Provides extra operation, analysis and
    visualization functionality for R package 'sigminer'.
}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
